move getMaxContainerHeight and getMaxContainerWidth to textElement.ts
